merge all current progress to print SMT2 models of compiled Cairo
Additionally, calls of functions are modelled in the CairoSemantics
     module by adding a 'pre → post' assertion. That required an
     additional Reader in the CairoSematnics monad. Since stockpiling
     of monad transformers gets a bit awkward to handle, the
     CairoSemantics was turned to a Free Monad DSL.

fix extracted 'memory' arguments to be wrapped in necessary 'let's

Actually, they are wrapped not just in necessary 'let's, but in all of
     the 'let's that the original expression had. Doesn't hurt,
     hopefully.

fix: de-UFify memory arguments — there can be nested memory accesses

fix: add jnz condition to pre-conditions for the following segments

remove condition from an arc from jnz to the fallthrough segment

The arc condition was negation of the jnz condition. That is, the arc
     condition asserts that the operand of jnz is zero. We can remove
     this condition for the following reason: the 'operand = 0'
     condition is already assumed on JNZs in the SMT encoder. By doing
     so, we gain the following: there are less segment breaks and
     instructions following the JNZ gain information about what
     preceded the JNZ.

stop clumping up usual assertions and pre-conditions

also, always add ap == fp to the list of funciton pre-conditions.

Before this change we treated pre-conditions, as if they were just
     assertions on the function label (effectively, on the first
     instruction of the function). That is wrong, because it
     necessitates satisfaction of the pre-conditions even on inner
     jumps (back to the first instruction), not just function calls.

always add a postcondition to rets, if it's omitted—use plain 'true'

Since having an assertion on a label serves as a module break, it must
     be present on all rets. Otherwise, a module won't have a finish
     and won't be added to the list of modules.

reject the assumption that each function must have at least one 'ret'