dirty: Remove eDSL from `CFGBuild.hs`.
This commit changes the type signature of the top-level
`solveContract` function called from `Main.hs`. In particular, this
function now has type:

        solveContract :: Env -> IO [SolvingInfo]

In order to do this, we completely overhaul the entire control flow
graph building procedure. The embedded DSL is removed, the monad stack
is gone, replaced with a simple `Either Text a`, and all functions are
either pure or live in `Either`. All code in `CFGBuild/Runner.hs` is
made irrelevant, and thus this source file is removed.

This commit leaves the `ContractInfo` type in a dirty state, as some
fields still have a type over `MonadError`. This also breaks the
`CairoSemantics` runner, which currently does not build.

* Remove unnecessary language extensions.
* Remove `CFGBuildF`, `CFGBuildL`.
* Move `CFG` to `CFGBuild.hs`.
* Move `Config` and `Env` to `ContractInfo.hs`.