Add manual profiling workflow, use newer bats-core
Bats-core 1.8.1 is now used, as it has a timing flag which bats-core 1.2 that was being used previously does not. Existing CI jobs are left unchanged, as we expect some of the tests run for profiling not to pass with certain solvers. For example, the tests using logical variables return an error when using mathsat. The profiling workflow runs each test from the testing job once for each solver, and outputs human readable results for each one including any errors encountered. Additionally a python script is used to convert these results into csv form such that any changes can be graphed relatively easily.