Print only a single `Unsat` judgement for all-`Unsat` functions.
Summary
=======
* stdout: e.g. `main+T` -> `main++T`
* stdout: Add newlines between judgements
* stdout: Sort output in special cases
* stdout: use `++` instead of `+` to indicate modules in output
* test: fix tests to accommodate stdout changes
* test: add `cvc5`, lower timeout, run `mathsat` before `z3`
* test: add `cvc5` installation to CI workflow
* test: use larger actions runner
* test: print number of processors in tests
* test: add `oneoff.sh` script for running a single test at a time

This commit adds some minor changes to `Global.hs` to group modules by
function name, and collapse them into a single judgement if all modules
for that function are `Unsat`. The test gold files are changed
accordingly. The changes of interest are in `solveContract`, in
particular, the line:

  +  pure $ (concat . map collapseAllUnsats . groupBy sameFuncName) infos

is added, along with the functions `collapseAllUnsats` and
`sameFuncName`.

To facilitate this change, we add an `si_funcName` field to the
`SolvingInfo` record type.

This commit also reformats the stdout of the tool to (1) change the
separating character between the function name and the module identifier
from `+` to `++`, as in:

old:
  main+T

new:
  main++T

This is done because in order to recover the function name for
collapsing `Unsat` outputs for modules of the same function, we split on
the separating character. However, a '+' sometimes appears in SMT
queries, which are dumped to stdout when segments are empty.

We also add newlines between the results for each module/function, which
makes the output a bit more readable.

Additionally, a bug is fixed where we did not properly collapse the
modules of `succpred` in the `func_multiple_ret.cairo` test. This is
fixed by sorting the `SolvingInfo`s, which necessitated writing an `Ord`
instance that only pays attention to the case where a function name is a
prefix of another function name.

Several other minor changes are made:
* Rename `inlinable` to `inlinables` to better reflect that this is a
  `Set ScopedFunction`.
* Change `makeModules` s.t. it takes a single `(CFG, ScopedFunction ->
  Bool)` argument, since its only use is uncurried.
* Add function `getModuleFuncName`, which splits on the '+' character.
* Construct `cfgs` s.t. it is a `[CFG]` instead of a list of tuples of
  CFGs and predicates. It is notable that since we zip `cfgs` and `map
  (==) fs`, we now have a data invariant that `cfgs` and `fs` are in the
  same order, and if we do any preprocessing of either of these that
  permutes that order, it could cause a bug.
* Derive `Eq` for `SolverResult`.

Separately, we also optimize the test invocation of `horus-check` to
dramatically reduce running time. By adding a `-s cvc5` and swapping the
order of `mathsat` and `z3` (which was already recently swapped, so we
are swapping it back), and dropping the timeout from 360000ms to
100000ms (since cvc5 is fast), we bring the total running time of the
tests on an 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz (4 cores, 8
threads) from ~19:00min to ~2:40min.