Add a hint about interpreting `Unknown` results.
This commit adds a hint that gets printed when we have at least one
`Unknown` in the output. The hint describes what `Unknown` means and how
to increase the timeout.

* Add `hie.yaml` for HLS.
* Use `unless` to simplify printing of `Unknown` hint in `app/Main.hs`.
* Add hint text to gold files with `Unknown`s.
* Refactor tests file `bats-template` to reduce duplication.

Since we now print hints, we must make the expected test output reflect
that, fixing some minor issues with the test suite. In particular, we
split the diffing in `bats-template` into a separate function.

We also add a couple `sed` calls to strip lines starting with `hint:`
from the Horus output, so that we don't get any false positives.

The profiling functions are collapsed down into a single function called
`single_solver_test_case` that takes an extra argument for the solver.

* Modify CI profiling workflow accordingly.
* Rename `entrypoint-profiling.sh` -> `profiling-entrypoint.sh`.