Minor refactor, collapse `ConstraintsState` and `ExecutionState` types.
This commit addresses a number of the comments from PR #84. We now have
just a single `ConstraintsState` record type, with an additional field
for the call stack.

We also refactor the `Ret` case in `addArcsFrom`, pulling some
functionality into helper functions.

The `solveContract` is also refactored slightly so that the types of
`cfg` and `cfgs` agree. We also simplify the logic for removing trusted
modules.

Other small changes:
* Remove unused functions
* Delete some TODOs
* Add comments in some places
* Rename `inlinable` to `inlinables`