Add support for arbitrary field primes parsed from JSON.
There is a key in the `"program"` field of the contract definition JSON
called `"prime"`, which is the characteristic of the finite field over
which all arithmetic is done in the contract. Horus used to hardcode
this to a value defined in `Util.hs` called `fieldPrime`, and even
though we parsed the actual value from the JSON, it was not used
anywhere.

This commit adds support for using the prime passed from `horus-compile`
via the JSON contract.
* Add a `GetProgram` constructor to `GlobalF`.
* Add an integer `fPrime` parameter to many functions.
* Refactor `outputSmtQueries`.
* Lint source files (remove a couple redundant brackets).