Rewrite `nameOfModule` to return separate parts of the name
This commit addresses a comment in the review of PR #89. In particular,
we refactor the aforementioned function into `getModuleNameParts`, which
returns a triple consisting of the prefix, labelsDigest, and oracle
suffix.

The prefix is the list of scopes for the given function, delimited by
`.` characters. The labelsDigest is a text representation of the labels
(there is usually just one). And the oracle suffix is an identifier for
the branch of control flow treated by the given module. For example, if
we have an `if-then-else`, the case where the predicate is true may have
an oracle suffix of `T`, and the case where the predicate is false may
have an oracle suffix of `F`. These would each be their own modules.

* Fix typo of `because` in hint.
* Remove `getModuleFuncName`.