Don't print separate results for labels in a given function
This commit addresses comments raised during the review of PR #89, which
implements a collection of changes to the stdout of the tool intended to
make it more friendly.

In particular, we address the concern raised by @aemartinez in
https://github.com/NethermindEth/horus-checker/pull/89#discussion_r1068516345.

We now combine results for labels inside functions with the results of
the functions themselves (but *only* in the case where everything is
`Verified`).

To achieve this, we add functionality to distinguish between floating
labels, like an `add:` inside a function, from function labels, like a
top-level function called `max`. Previously, these top-level names were
included in the `labelsDigest` (now called `labelsSummary`).

At a high-level, a great deal of the logic for constructing module names
has been rewritten, in particular `normalizedName`. We also add detailed
docstrings for all touched functions.

Some terminology is changed. In particular:
* `prefix` -> `qualifiedFuncName`
* `labelsDigest` -> `labelsSummary`
* `funcNamePrefix` -> `labeledFuncName`

We no longer drop the first part of a `ScopedName` for printing, which
means the trusted specs are adjusted. However, we still drop `__main__`
to avoid cluttering the output.

Other changes:
* Patch gold files for relevant tests.
* Format source files with `fourmolu`.