Rewrite `Unknown` hint
When a user gets a judgement of Unknown, we print a hint explaining what
this means, as well as a suggestion to increase the timeout. In this
hint, there is an example that explains the somewhat unintuitive fact
that the timeout setting is *not* the maximum running time of a call to
`horus-check`, but rather the maximum running time of each SMT query, of
which there may be several per-function.

This commit better explains the difference between query count and
function count, gives a true upper bound for running time in the
example, and explains *why* there are sometimes multiple queries for a
single function.