Move `inlinable` boolean inside `SolvingInfo` sum type
This commit is a tiny refactor of the logic introduced by PR #103, which
added the `[inlined]` mark after function names that are inlinable in
the output of the tool. Instead of carrying around a 2-tuple everywhere,
we put the `inlinable` status of the module inside the `SolvingInfo`
type as a fourth field.