Add `:::default` oracle suffixes when we don't collapse results
This commit addresses some confusing behavior in the output for functions with
multiple modules where at least one module is `False`.

Let’s take `func_gauss.cairo` as example, the `gauss()` function has a loop
labeled `loop` with it’s corresponding invariant. As everything is correct we
get

```console
gauss
Verified

main
Verified
```

But if we break the code inside the loop as to invalidate the invariant, we get
something like this:

```console
gauss
Verified

gauss.loop:::T
Verified

gauss.loop:::F
False

main
Verified
```

The verifier is successfully identifying there’s a problem inside the loop, but
we still get `gauss Verified`. The reason is because in the first case `gauss
Verified` means the whole function is correct (because everything about
`gauss()` is packed in one line as everything is `Verified`), but in the second
case, `gauss Verified` just means "the module before the loop label has been
checked".

To remedy this issue, we now print `gauss:::default` in the second case.

* Add `:::default` oracle suffixes in test gold files