Break module name formatting into two parts
This commit replaces the `normalizedName` function with several new
functions, removing use of partial helper functions like `init` and
`head`, and instead making use of `Data.List.NonEmpty`. The new
functions are:

* preprocessScopes
* formatScopes
* summarizeLabels

Additionally, we rename:
* summarizeLabels -> summarizeLabels'
* sansCommonAncestor -> detachCommonPrefix

This fixes a bug where a dot is not added between function name and
label for anonymous asserts.

We no longer generate the `prefix` and `labelsSummary` within the same
function, because that was kind of arbitrary.

Other things:

* Adjust gold files to remove assert label, inlined markers
* Add test `assert_let_minimal.cairo`
* Fix order of results in `assert_let.gold`