Add working prefix of `vat.cairo` (via manually flattnened storage vars)
This commit removes the porting comments in `vat.cairo` because they
were crowding the source code. We flatten the storage vars manually by
literally just replacing a single struct-valued storage var with n
storage vars, where n is the number of fields of the struct.

We also add valid `@storage_update` annotations for this prefix of the
felt-converted `frob()` function.