dirty: Add another set of asserts to `frob()`
This commits adds some more lines of the old `frob()` function to the
specified version. We also use `stack run` in the frobenius display
script, and add the ability to whitelist modules to avoid waiting for
extraneous stuff to solve.

This debugging code is dangerous and should be removed. It only checks
modules whose name starts with `frob`.