Fix improper handling of post with a reference to storage
This commit fixes a bug where storage variables were dereferenced with
respect to the state of the `ConstraintsState` `Storage` mapping *after*
the function being spec-ed had been processed, and in particular, taking
into account any `.write()` calls within the function body.

The expected behavior is that references to storage variables within
postconditions refer to the state of the storage variable immediately
prior to the call of the function being specified.

In simple terms, the value of a storage variable in an annotation is
always the value *before* any storage writes in the current function.

The 'exception' is the LHS of `@storage_update` annotations themselves,
but this is more of an assignment than a reference.

* Remove unnecessary checks for accessing storage in plain specs.
  Basically, before, it looks as though we didn't allow storage reads
  within plain specs, which is wrong. After all, a plain spec is
  more-or-less defined to be a spec for a function without a 'write'.

* Change readStorage to allow for manual override of the point in time
  in which one reads from a storage variable. Implementation-wise, we
  allow you to pass a value of type `Storage` to read with respect to.

* Use the new `readStorage` in mkCallConstraints` to make sure that when
  you refer to a storage variable in a postcondition, you are referring
  to its state prior to the function being called.

Other minor changes:
* Add docstrings for:
  - `encodeRichSpec`
  - `encodePlainSpec`
  - `storageRemoval'`
  - `mkInstructionConstraints`
* Add a test for the above bug: `post_with_svar.cairo`
* Format sources with `fourmolu`