Remove `Checkpoint`, check `@pre`s in separate modules
This commit sees a heavy refactor of the control-flow graph building
procedure, and the `CairoSemantics` logic. The most notable thing is
removing the `CheckPoint` interpreter branch of the `CairoSemantics`
eDSL, which was formerly used to handle existentials, and adding
so-called "optimising" modules, which are modules used solely to check
preconditions of called functions within some parent module.

We also add a `Vertex` type used to construct CFGs, which allows us to
distinguish between vertices and plain-old labels.

Additionally, we refactor module name construction to include a
serialization of the call stack.

Other notable changes
=====================
* Add `Show` instance for `Expr a`
* Add `optimisesF` field to `Module` and `SolvingInfo` types
* Report a collapsed result as inlinable if and only if all
  `SolvingInfo`s are inlinable
* Remove `hasStorage` check when detecting inlinable functions
* Modify `*.gold` files for many tests to reflect new module naming
* Add new test `optimise_notsplit_many.cairo`
* Add new test `storage_pre.cairo`