Add refactored `storageVarsSpecs`
This commit makes changes to match the new `spec.json` format in:
https://github.com/NethermindEth/horus-compile/pull/63

In particular, return types of storage variables which are `struct`s are
flattened (deeply) into a tuple, and internally, we use an extra argument
to index into this tuple.

Parsing the new JSON format
===========================
For example, the keys in the `storage_update` object in the `spec.json`
file have been modified as follows:

```diff
- 	"__main__.complex_var_1.s.x": [
+ 	"__main__.complex_var_1 0": [
```

The relevants bits of the source file are:

```cairo
struct Point {
    x: felt,
    y: felt,
}

@storage_var
func complex_var_1(arg: Uint256) -> (s: Point) {
}
```

Note that instead of using dot-accessor notation, we use an index `0` to
indicate the first member of  the  return  tuple  of  `complex_var_1()`,
which is identified with the field `x` of `Point`.

In order to parse this index argument from the key in the JSON, as  seen
in the diff above, we  implicitly  change  the  `FromJSON`  instance  of
`FuncSpec`/`Storage`,  and  define   `StorageUpdateKey`   which   is
basically `"__main__.complex_var_1 0"`, as is  obvious  from  its  name.

So instead of parsing a value of type `Storage` from the JSON, we  parse
a value of type `IndexedStorage`:

```
+type IndexedStorage = Map StorageUpdateKey [([Expr TFelt], Expr TFelt)]
```

and then convert it to a map of type `Storage` with a `unindexStorage`
function.

Standard specs for `<storage>.read()` and `<storage>.write()`
=============================================================
We refactor `storageVarsSpecs` heavily, folding over the
`ContractDefinition` field:

```diff
+  , cd_storageVars :: Map ScopedName (Arity, Coarity)
```

which now includes the arity and coarity of each storage variable.

In order to manage this, we add helper functions:

```diff
+mkReadSpecs :: ScopedName -> (Int, Int) -> [(ScopedName, FuncSpec)]
+mkWriteSpecs :: ScopedName -> (Int, Int) -> [(ScopedName, FuncSpec)]
```

which basically construct a separate `FuncSpec` for each element of  the
flattened return tuple.  We use a `mergeFuncSpecs`  function  to  nicely
take the conjunction of the preconditions and postconditions,  and  take
the union of the `Storage` maps.