Eliminate `RichSpec`, `ModuleData` types
This commit uses `FuncSpec` for everything. A `PlainSpec` is now just a
`FuncSpec` with an empty `Storage` map. As a result, we no longer have a
distinction between rich specs and plain specs. Now everything is
handled by `encodeModule`, which is just `encodeRichSpec` with the logic
of `encodePlainSpec` inlined.

We remove all of the following:
* `ModuleData` (this was a bad idea in current state, sorry!)
* `ModuleSpec`
* `PlainSpec`
* `RichSpec`